Fix mosaic stop dev-origin
