+++ src/bin/mosaic-attach.ts
@@ -5,13 +5,13 @@ import Shell from '../Shell';
 import Node from '../Node/Node';
 import NodeFactory from '../Node/NodeFactory';
 import NodeDescription from '../Node/NodeDescription';
-import DevChainOptions from "./DevChainOptions";
+import ChainInfo from "../Node/ChainInfo";
 
 mosaic
   .arguments('<chain>')
   // Chain can't be validated as origin chain id is not received for aux chain.
   .action((chain: string) => {
-    const chainIdentifier = DevChainOptions.getDevChainParams(chain).chain;
+    const chainIdentifier = ChainInfo.getChainParams(chain).chain;
     const node: Node = NodeFactory.create(new NodeDescription(chainIdentifier));
 
     const args = [
