+++ src/bin/mosaic-logs.ts
@@ -5,12 +5,12 @@ import Shell from '../Shell';
 import Node from '../Node/Node';
 import NodeDescription from '../Node/NodeDescription';
 import NodeFactory from '../Node/NodeFactory';
-import DevChainOptions from "./DevChainOptions";
+import ChainInfo from "../Node/ChainInfo";
 
 mosaic
   .arguments('<chain>')
   .action((chain: string) => {
-    const chainIdentifier = DevChainOptions.getDevChainParams(chain).chain;
+    const chainIdentifier = ChainInfo.getChainParams(chain).chain;
     // Chain can't be validated as origin chain id is not received for aux chain.
     const node: Node = NodeFactory.create(new NodeDescription(chainIdentifier));
     const args = [
